Move snackbar to root level
- still causing memory leak
